# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stars Decoded Natal Chart Report API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List the viewer's people (profiles)
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileSummary"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create or de-dupe-resolve a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Profile created or matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileSummary"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profiles/{id}:
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update mutable profile fields
      description: >
        Allows the owner to update mutable profile fields.
        Currently supports toggling `isSelf`. Setting `isSelf=true` automatically
        clears any other self-profile for the same user (one per user invariant).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileSummary"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the owner of this profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profiles/{id}/birth-time:
    patch:
      operationId: updateProfileBirthTime
      tags: [profiles, horizon]
      summary: Add or correct the birth time and run the horizon pass
      description: >
        Owner only. Stores the new time and window on the profile, recomputes
        its chart, and runs a horizon pass on every complete report of the
        profile: the previous text is kept in report_revisions, the horizon
        sections are generated and the rest is amended by quote match, never
        regenerated (ADR-35). The first pass is free (MB-52).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBirthTimeBody"
      responses:
        "202":
          description: The pass has started on the listed reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BirthTimeUpdateResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found or not owned by the viewer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A pass is already running on one of the profile's reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    UpdateBirthTimeBody:
      type: object
      properties:
        birthTime:
          type: string
        birthTimeWindowMinutes:
          type: integer
      required: [birthTime, birthTimeWindowMinutes]
    BirthTimeUpdateResponse:
      type: object
      properties:
        profileId:
          type: string
        horizon:
          type: string
          enum: [known, approximate, unknown]
        reportIds:
          type: array
          description: The reports a pass has started on.
          items:
            type: string
      required: [profileId, horizon, reportIds]
...
    ProfileSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        birthDate: { type: string }
        birthTime: { type: string }
        birthPlace: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        timezoneOffset: { type: number }
        timezone: { type: string, nullable: true }
        birthTimeWindowMinutes: { type: integer }
        horizon:
          type: string
          enum: [known, approximate, unknown]
          nullable: true
        sunSign: { type: string, nullable: true }
        moonSign: { type: string, nullable: true }
        risingSign: { type: string, nullable: true }
        createdAt: { type: string }
        ownership:
          type: string
          enum: [owner, claimed, invited, unclaimed]
          description: Viewer-relative ownership status of the profile.
        claimedByName: { type: string, nullable: true }
        inviteEmail: { type: string, nullable: true }
        isSelf:
          type: boolean
          description: >
            True if this profile is the authenticated viewer's own self-profile.
            Set explicitly when the user created the report via the
            "Generate My Chart" flow (isForSelf=true in POST /reports).
            Never derived by heuristic. Always false for anonymous viewers.
      required: [id, name, birthDate, birthTime, birthPlace, createdAt]
    CreateProfileBody:
      type: object
      properties:
        name: { type: string }
        birthDate: { type: string }
        birthTime: { type: string }
        birthPlace: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        timezoneOffset: { type: number }
        timezone: { type: string, description: "IANA zone name; preferred over timezoneOffset (MB-48)." }
        birthTimeWindowMinutes: { type: integer, description: "0 exact, 180 a part of the day, 720 unknown (ADR-33). Default 0." }
      required: [name, birthDate, birthTime, birthPlace, latitude, longitude, timezoneOffset]
    UpdateProfileBody:
      type: object
      description: Mutable profile fields the owner can update.
      properties:
        isSelf:
          type: boolean
          description: >
            Set to true to designate this profile as the viewer's own self-profile.
            Setting true automatically clears isSelf on all other profiles owned by
            the same user (one per user). Set to false to unmark it.
...